Finally added core functionality to notify attendant of low stocks. Using 20% of the post restock amount.

+++ store/main/views.py
@@ -1,4 +1,4 @@
-from flask import render_template, url_for, redirect, request, flash, current_app, send_from_directory
+from flask import render_template, url_for, redirect, request, flash, current_app
 from . import main
 from .. import db
 from .forms import CreateItem, RestockItems
@@ -12,9 +12,14 @@ def favicon():
     return current_app.send_static_file('favicon.ico')
 
 
+def filter_low_stock(items):
+    return [item for item in items if item.low_stock]
+
+
 @main.route('/', methods=['GET', 'POST'])
 def index():
     items = Item.query.all()
+    low_stock_items = filter_low_stock(items)
 
     if request.method == 'POST':
         # take the form's ID and quantity fields
@@ -24,8 +29,11 @@ def index():
         # query the item
         item = Item.query.get(id)
         # prevent invalid entries
-        if quantity <= 0 or quantity > item.stock:
-            flash('Invalid purchase quantity!', 'error')
+        if quantity == 0:
+            flash('ℹ Please enter a valid purchase amount', 'info')
+            return redirect(url_for('.index'))
+        if quantity < 0 or quantity > item.stock:
+            flash('Invalid purchase quantity ❌', 'error')
             return redirect(url_for('.index'))
 
         # create cart item from above data and add to DB session and commit
@@ -35,8 +43,11 @@ def index():
                             sub_total = quantity * item.price)
         db.session.add(cart_row)
         db.session.commit()
-        flash('{} added to cart.'.format(item.name), 'info')
+        flash('ℹ {} added to cart'.format(item.name), 'info')
         return redirect(url_for('.index'))
+    
+    if low_stock_items:
+        flash('⚠ Some items are low on stock. Please visit the restock page to see them ⚠', 'warning')
     return render_template('main/home.html', items = items)
 
 
@@ -47,35 +58,42 @@ def cart():
 
     if request.method == 'POST':
         row_ids = request.form.getlist('remove_item')
-        # if no items were selected for deletion, proceed to checkout logic
-        if row_ids == []:
-            """Confirm checkout logic. 
-            
-            Basically applies purchase quantities and updates item stock counts, then proceeds to empty cart table"""
-            for row in cart_rows:
-                # query respective items
-                item = Item.query.get(row.id)
-                # apply purchase to stock
-                item.stock -= row.purchase_quantity
-                if item.stock == 0:
-                    item.sold_out == True
-                # add item back
-                db.session.add(item)
-            # delete all cart row items
-            for row in cart_rows:
-                db.session.delete(row)
-            # commit session and redirect
-            db.session.commit()
-            flash('Checkout Complete! Return to Store Front.', 'success')
-            return redirect(url_for('.cart'))
 
-        """Delete items from cart"""
-        # loop though list and delete each item from DB
-        for row in row_ids:
-            cart_row = CartRow.query.get(int(row))
-            db.session.delete(cart_row)
-        db.session.commit()
-        return redirect(url_for('.cart'))
+        # if no items are selected for deletion
+        if row_ids == []:
+            # if cart table contains items, proceed to checkout
+            if cart_rows != []:
+                """Confirm checkout logic. 
+                
+                Basically applies purchase quantities and updates item stock counts, then proceeds to empty cart table"""
+                for row in cart_rows:
+                    # query respective items
+                    item = Item.query.get(row.id)
+                    # apply purchase to stock
+                    item.stock -= row.purchase_quantity
+                    item.verify_stock_status()
+                    # update item in DB
+                    db.session.add(item)
+                
+                # empty cart table
+                for row in cart_rows:
+                    db.session.delete(row)
+                # commit session and redirect
+                db.session.commit()
+                flash('Checkout Complete! Return to Store Front.', 'success')
+                return redirect(url_for('.cart'))
+            else:
+                flash('Cart is empty 🛒✖', 'error')
+                return redirect(url_for('.cart'))
+        else:
+            """Delete items from cart"""
+            # loop though list and delete each item from DB
+            for row in row_ids:
+                cart_row = CartRow.query.get(int(row))
+                db.session.delete(cart_row)
+            db.session.commit()
+            return redirect(url_for('.cart'))
+        
     return render_template('main/cart.html', cart_rows = cart_rows, cart_total = cart_total)
 
 
@@ -99,9 +117,10 @@ def create_item():
                     price = create.price.data,
                     stock = create.stock.data,
                     post_restock = create.stock.data)
+        item.verify_stock_status()
         db.session.add(item)
         db.session.commit()
-        flash('Item successfully added!', 'success')
+        flash('Item successfully added ✔', 'success')
         return redirect(url_for('.create_item'))
     return render_template('main/create.html', create = create)
 
@@ -109,6 +128,7 @@ def create_item():
 @main.route('/restock', methods=['GET', 'POST'])
 def restock():
     items = Item.query.all()
+    low_stock_items = filter_low_stock(items)
     item_names = [x.name for x in items]
     modded_item = None
     restock_form = RestockItems()
@@ -119,7 +139,7 @@ def restock():
 
         # ensure user does not restock a negative number
         if addition < 1 or name not in item_names:
-            flash('Sorry, you have entered an invalid stock amount.', 'error')
+            flash('Sorry, you have entered an invalid stock amount ❌', 'error')
             return redirect(url_for('.restock'))
 
         for item in items:
@@ -127,18 +147,21 @@ def restock():
                 modded_item = item
                 break
 
-        modded_item.stock += addition
-        modded_item.post_restock = modded_item.stock
+        modded_item.restock(addition)
         db.session.add(modded_item)
         db.session.commit()
-        flash('{}+ {} added to stock'.format(addition, item.name), 'info')
+        flash('ℹ {} {} added to stock'.format(addition, item.name), 'info')
         return redirect(url_for('.restock'))
-    return render_template('main/restock.html', items = items, restock_form = restock_form)
+
+    if low_stock_items:
+        flash('⚠ Kindly restock the following items:', 'warning')
+    return render_template('main/restock.html', items = items, restock_form = restock_form, low_stock_items = low_stock_items)
 
 
 @main.route('/manage_items', methods=['GET', 'POST'])
 def manage_items():
     items = Item.query.all()
+    low_stock_items = filter_low_stock(items)
 
     if request.method == 'POST':
         # list of checked boxes
@@ -146,11 +169,15 @@ def manage_items():
         if ids == []:
             flash('Please select at least one item.', 'info')
             return redirect(url_for('.manage_items'))
+
         to_be_removed = [Item.query.get(id) for id in ids]
 
         for item in to_be_removed:
             db.session.delete(item)
         db.session.commit()
-        flash('Item(s) removed.', 'success')
+        flash('Item(s) removed ✔', 'success')
         return redirect(url_for('.manage_items'))
+
+    if low_stock_items:
+        flash('⚠ Some items are low on stock. Please visit the restock page to see them ⚠', 'warning')
     return render_template('main/manage.html', items = items)
